ci(workflows): Add pypi-publish.yml, remove python-publish.yml
Added a new workflow file for publishing to PyPI and removed the old workflow file for publishing Python packages. The new workflow uses PDM for Python setup and package building, and it publishes the package distributions to PyPI using the pypa/gh-action-pypi-publish action. Also updated the project version in pyproject.toml to 1.0.1.